pyproject.toml
[project]
name = "kartvision"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "わんぱん", email = "zhiyingxiqi@gmail.com" }]
dependencies = [
    "pyautogui>=0.9.54",
    "google-cloud-vision>=3.7.4",
    "flask>=3.0.3",
    "pillow>=10.4.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/kartvision"]

[tool.rye.scripts]
kart = { cmd = "flask --app src/kartvision/app.py run" }
